Major refactor, putting pile logic into game_state
Much of the logic that was in the pile class was
beginning to seem out of place. A lot of the rules
have fairly arbitrary interactons *between* piles,
which weren't best represented in the pile classes
themselves, which were getting unweildy and
inefficient. Hence, much of this logic has been
moved to the game state class.