Fix the build with recent GCCs.
It failed to build on Mageia Linux v8 x86-64 with:

<<<
gcc (Mageia 10.2.0-2.mga8) 10.2.0
Copyright (C) 2020 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
>>>

The changes should be backward compatible because they only add missing
includes.